refactor: 운동거리 int로 변경

+++ src/main/java/clofi/runningplanet/running/dto/RunningStatusResponse.java
@@ -7,7 +7,7 @@ public record RunningStatusResponse(
 	Long memberId,
 	String nickname,
 	int runTime,
-	double runDistance,
+	int runDistance,
 	boolean isEnd
 ) {
 	public RunningStatusResponse(Member member, Record savedRecord) {
